Update test script

+++ Scripts/ble_test.py
@@ -48,7 +48,7 @@ async def write_file(address, offset, local_filename, remote_filename):
         # Sending packet to signal completion
         await client.write_gatt_char(CRS_RX_UUID, b'\x10')
 
-async def read_file(address, offset, remote_filename, local_filename, timeout=1):
+async def read_file(address, offset, stride, remote_filename, local_filename, timeout=1):
     with tqdm(desc="Receiving Bytes", unit="B", unit_scale=True) as pbar:
         async with BleakClient(address, adapter='hci0') as client:
             file_data = bytearray()
@@ -67,8 +67,9 @@ async def read_file(address, offset, remote_filename, local_filename, timeout=1)
                     packet_received.set()
 
             await client.start_notify(CRS_TX_UUID, file_notification_handler)
-            offset_bytes = offset.to_bytes(4, byteorder='little')
-            await client.write_gatt_char(CRS_RX_UUID, b'\x02' + offset_bytes + remote_filename.encode())
+            offset_bytes = offset.to_bytes(2, byteorder='little')
+            stride_bytes = stride.to_bytes(2, byteorder='little')
+            await client.write_gatt_char(CRS_RX_UUID, b'\x02' + offset_bytes + stride_bytes + remote_filename.encode())
 
             try:
                 while not transfer_complete.is_set():
@@ -142,8 +143,8 @@ def main():
     parser.add_argument('--list', action='store_true', help='List all available BLE devices.')
     parser.add_argument('--address', type=str, metavar='ADDRESS', help='Specify the BLE device address.')
     parser.add_argument('--dir', type=str, metavar='DIRECTORY', help='List contents of the specified directory.')
-    parser.add_argument('--read', nargs='+', help='Read a file from the device. Provide the offset, remote file path, and an optional local file path.')
-    parser.add_argument('--write', nargs=3, metavar=('OFFSET', 'LOCAL_FILE', 'REMOTE_FILE'), help='Write a file to the device. Provide the offset, local file path, and remote file path.')
+    parser.add_argument('--read', nargs='+', help='Read a file from the device. Provide the offset, stride, remote file path, and an optional local file path.')
+    parser.add_argument('--write', nargs=2, metavar=('LOCAL_FILE', 'REMOTE_FILE'), help='Write a file to the device. Provide the local file path, and remote file path.')
     parser.add_argument('--create', type=str, metavar='FILE_NAME', help='Create a new file on the device with the specified name.')
     parser.add_argument('--delete', type=str, metavar='FILE_NAME', help='Delete the specified file from the device.')
     parser.add_argument('--mkdir', type=str, metavar='DIRECTORY_NAME', help='Create a new directory on the device with the specified name.')
@@ -153,18 +154,18 @@ def main():
         asyncio.run(list_devices())
     elif args.address and args.dir:
         asyncio.run(list_directory(args.address, args.dir))
-    elif args.read:
-        offset, remote_filename, *local_filename = args.read
+    elif args.address and args.read:
+        offset, stride, remote_filename, *local_filename = args.read
         local_filename = local_filename[0] if local_filename else os.path.basename(remote_filename)
-        asyncio.run(read_file(args.address, int(offset), remote_filename, local_filename))
-    elif args.write:
+        asyncio.run(read_file(args.address, int(offset), int(stride), remote_filename, local_filename))
+    elif args.address and args.write:
         offset, local_filename, remote_filename = args.write
         asyncio.run(write_file(args.address, int(offset), local_filename, remote_filename))
-    elif args.create:
+    elif args.address and args.create:
         asyncio.run(create_file(args.address, args.create))
-    elif args.delete:
+    elif args.address and args.delete:
         asyncio.run(delete_file(args.address, args.delete))
-    elif args.mkdir:
+    elif args.address and args.mkdir:
         asyncio.run(mkdir(args.address, args.mkdir))
     else:
         print("Invalid arguments. Use --help for usage information.")
